changed binarySearchBefore to findInterval moved tests, added test giving relation to binarySearch
